# nodia-server/test/fixtures/finance-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/finance/category-groups": {
      "get": {
        "operationId": "FinanceCategoryGroupController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "active",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "all"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Bounded Ransack filters; no user_id or is_active.",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "category_ids",
            "required": false,
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "category_group_ids",
            "required": false,
            "in": "query",
            "schema": {
              "maxItems": 100,
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "obligation_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "user_id",
                          "name",
                          "key",
                          "is_active",
                          "created_at",
                          "updated_at",
                          "category_count"
                        ],
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string",
                            "pattern": "^[1-9]\\d{0,18}$",
                            "example": "42"
                          },
                          "user_id": {
                            "type": "string",
                            "pattern": "^[1-9][0-9]*$",
                            "example": "1"
                          },
                          "is_active": {
                            "type": "boolean"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          },
                          "key": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          },
                          "category_count": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 100
                          }
                        }
                      },
                      "maxItems": 100
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "page",
                        "limit",
                        "total_items",
                        "total_pages"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "page": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 1000000
                        },
                        "limit": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 100
                        },
                        "total_items": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "total_pages": {
                          "type": "integer",
                          "minimum": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, query, bigint ID, or category selection.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "A validated session is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group or associated category does not exist in the authenticated user scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate user-scoped key, new inactive category selection, or association conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List personal category groups with bounded pagination",
        "tags": [
          "Finance category groups"
        ]
      },
      "post": {
        "operationId": "FinanceCategoryGroupController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFinanceCategoryGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "user_id",
                    "name",
                    "key",
                    "is_active",
                    "created_at",
                    "updated_at",
                    "category_count",
                    "categories"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^[1-9]\\d{0,18}$",
                      "example": "42"
                    },
                    "user_id": {
                      "type": "string",
                      "pattern": "^[1-9][0-9]*$",
                      "example": "1"
                    },
                    "is_active": {
                      "type": "boolean"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "key": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "category_count": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 100
                    },
                    "categories": {
                      "type": "array",
                      "maxItems": 100,
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "id",
                          "name",
                          "key",
                          "is_active"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "pattern": "^[1-9]\\d{0,18}$",
                            "example": "42"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          },
                          "key": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 255
                          },
                          "is_active": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input, query, bigint ID, or category selection.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "A validated session is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group or associated category does not exist in the authenticated user scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Duplicate user-scoped key, new inactive category selection, or association conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "timestamp",
                    "path",
                    "method",
                    "error",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "path": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create a personal category group and its memberships atomically",
        "tags": [
          "Finance category groups"
        ]
      }
    },
...
  },
  "info": {
    "title": "Personal Finance isolated contract",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreateFinanceCategoryGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255
          },
          "key": {
            "type": "string",
            "maxLength": 255
          },
          "category_ids": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "is_active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name",
          "key",
          "category_ids"
        ]
      },
...
    }
  }
}